Restructure commands under more specific click groups
