+++ .github/workflows/pythonpackage.yml
@@ -30,11 +30,11 @@ jobs:
 
     - name: Python License Check
       run: |
-        python -m detection_rules license-check
+        python -m detection_rules dev license-check
         
     - name: Build release package
       run: |
-        python -m detection_rules build-release
+        python -m detection_rules dev build-release
 
     - name: Archive production artifacts
       uses: actions/upload-artifact@v2
