+++ detection_rules/__init__.py
@@ -3,8 +3,10 @@
 # you may not use this file except in compliance with the Elastic License.
 
 """Detection rules."""
+from . import devtools
 from . import docs
 from . import eswrap
+from . import kbwrap
 from . import main
 from . import mappings
 from . import misc
@@ -14,8 +16,10 @@ from . import schemas
 from . import utils
 
 __all__ = (
+    'devtools',
     'docs',
     'eswrap',
+    'kbwrap',
     'mappings',
     "main",
     'misc',
