+++ detection_rules/utils.py
@@ -232,3 +232,23 @@ def load_rule_contents(rule_file: str, single_only=False) -> list:
             return rule
         else:
             raise ValueError(f"Expected a list or dictionary in {rule_file}")
+
+
+def format_command_options(ctx):
+    """Echo options for a click command."""
+    formatter = ctx.make_formatter()
+    opts = []
+
+    for param in ctx.command.get_params(ctx):
+        if param.name == 'help':
+            continue
+
+        rv = param.get_help_record(ctx)
+        if rv is not None:
+            opts.append(rv)
+
+    if opts:
+        with formatter.section('Options'):
+            formatter.write_dl(opts)
+
+    return formatter.getvalue()
